+++ src/main/java/org/folio/entlinks/client/InstanceStorageClient.java
@@ -13,7 +13,7 @@ public interface InstanceStorageClient {
   @GetMapping(value = "/instances", produces = APPLICATION_JSON_VALUE)
   InventoryInstanceDtoCollection getInstanceStorageInstances(@RequestParam String query, @RequestParam int limit);
 
-  record InventoryInstanceDto(String id, String title) { }
+  record InventoryInstanceDto(String id, String title, String source) { }
 
   record InventoryInstanceDtoCollection(List<InventoryInstanceDto> instances) { }
 }