+++ src/main/java/org/folio/entlinks/controller/delegate/InstanceAuthorityStatServiceDelegate.java
@@ -6,6 +6,7 @@ import static org.folio.entlinks.utils.DateUtils.fromTimestamp;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -15,8 +16,10 @@ import org.folio.entlinks.domain.dto.AuthorityControlMetadata;
 import org.folio.entlinks.domain.dto.AuthorityStatsDto;
 import org.folio.entlinks.domain.dto.AuthorityStatsDtoCollection;
 import org.folio.entlinks.domain.dto.LinkAction;
+import org.folio.entlinks.domain.entity.AuthorityBase;
 import org.folio.entlinks.domain.entity.AuthorityDataStat;
 import org.folio.entlinks.domain.repository.AuthoritySourceFileRepository;
+import org.folio.entlinks.service.consortium.ConsortiumTenantsService;
 import org.folio.entlinks.service.links.AuthorityDataStatService;
 import org.folio.entlinks.utils.DateUtils;
 import org.folio.spring.client.UsersClient;
@@ -33,6 +36,7 @@ public class InstanceAuthorityStatServiceDelegate {
   private final DataStatsMapper dataStatMapper;
   private final UsersClient usersClient;
   private final AuthoritySourceFileRepository sourceFileRepository;
+  private final ConsortiumTenantsService tenantsService;
 
   public AuthorityStatsDtoCollection fetchAuthorityLinksStats(OffsetDateTime fromDate, OffsetDateTime toDate,
                                                               LinkAction action, Integer limit) {
@@ -47,6 +51,7 @@ public class InstanceAuthorityStatServiceDelegate {
     }
 
     var users = getUsers(dataStatList);
+    var isCentralTenant = tenantsService.isCentralTenantContext();
     var stats = dataStatList.stream()
       .map(source -> {
         var authorityDataStatDto = dataStatMapper.convertToDto(source);
@@ -54,6 +59,10 @@ public class InstanceAuthorityStatServiceDelegate {
         if (authorityDataStatDto != null) {
           fillSourceFiles(authorityDataStatDto);
           authorityDataStatDto.setMetadata(getMetadata(users, source));
+          var shared = isCentralTenant || Optional.ofNullable(source.getAuthority())
+            .map(AuthorityBase::isConsortiumShadowCopy)
+            .orElse(false);
+          authorityDataStatDto.setShared(shared);
         }
         return authorityDataStatDto;
       })