+++ src/main/java/org/folio/entlinks/controller/delegate/LinkingServiceDelegate.java
@@ -2,11 +2,11 @@ package org.folio.entlinks.controller.delegate;
 
 import static java.util.Objects.isNull;
 import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.folio.entlinks.utils.DateUtils.fromTimestamp;
 
 import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -27,6 +27,7 @@ import org.folio.entlinks.service.consortium.propagation.ConsortiumLinksPropagat
 import org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService;
 import org.folio.entlinks.service.consortium.propagation.model.LinksPropagationData;
 import org.folio.entlinks.service.links.InstanceAuthorityLinkingService;
+import org.folio.entlinks.utils.ConsortiumUtils;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.tenant.domain.dto.Parameter;
 import org.jetbrains.annotations.NotNull;
@@ -57,14 +58,13 @@ public class LinkingServiceDelegate {
     var links = linkingService.getLinks(status, fromDate, toDate, limit + 1);
     log.debug("Retrieved links count {}", links.size());
 
-    if (links.size() > limit) {
-      var nextDate = fromTimestamp(links.get(limit).getUpdatedAt());
-      bibStatsCollection.setNext(nextDate);
-      links = links.subList(0, limit);
-    }
-
     var stats = statsMapper.convertToDto(links);
-    fillInstanceTitles(stats);
+    stats = filterOutShadowCopiesAndFillInstanceTitles(stats);
+    if (stats.size() > limit) {
+      var nextDate = stats.get(limit).getUpdatedAt();
+      bibStatsCollection.setNext(nextDate);
+      stats = stats.subList(0, limit);
+    }
 
     return bibStatsCollection.stats(stats);
   }
@@ -123,25 +123,35 @@ public class LinkingServiceDelegate {
     }
   }
 
-  private void fillInstanceTitles(List<BibStatsDto> bibStatsList) {
+  private List<BibStatsDto> filterOutShadowCopiesAndFillInstanceTitles(List<BibStatsDto> bibStatsList) {
     var instanceIds = bibStatsList.stream()
       .map(BibStatsDto::getInstanceId)
       .map(UUID::toString)
       .distinct()
       .toList();
 
-    var instanceTitles = instanceService.getInstanceTitles(instanceIds);
+    var instanceData = instanceService.getInstanceData(instanceIds);
 
+    var bibStatsResult = new LinkedList<BibStatsDto>();
     bibStatsList.forEach(bibStatsDto -> {
       var instanceId = bibStatsDto.getInstanceId().toString();
-      var title = instanceTitles.get(instanceId);
+      var instanceDataEntry = instanceData.get(instanceId);
+      var isShadowCopy = instanceDataEntry == null
+        || ConsortiumUtils.isConsortiumShadowCopy(instanceDataEntry.getRight());
+      if (isShadowCopy) {
+        return;
+      }
 
+      var title = instanceDataEntry.getLeft();
       if (isBlank(title)) {
         log.warn("Title for instance {} is blank", instanceId);
         return;
       }
 
       bibStatsDto.setInstanceTitle(title);
+      bibStatsResult.add(bibStatsDto);
     });
+
+    return bibStatsResult;
   }
 }