+++ src/main/java/org/folio/entlinks/service/consortium/ConsortiumTenantsService.java
@@ -4,9 +4,11 @@ import static org.folio.entlinks.config.constants.CacheNames.CONSORTIUM_TENANTS_
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.entlinks.client.ConsortiumTenantsClient;
+import org.folio.spring.FolioExecutionContext;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,7 @@ public class ConsortiumTenantsService {
 
   private final UserTenantsService userTenantsService;
   private final ConsortiumTenantsClient consortiumTenantsClient;
+  private final FolioExecutionContext context;
 
   /**
    * Get consortium tenants for tenantId.
@@ -37,6 +40,15 @@ public class ConsortiumTenantsService {
       log.debug("Unexpected exception occurred while trying to get consortium tenants", e);
       return Collections.emptyList();
     }
+  }
+
+  /**
+   * Check if current context tenant is a consortium central tenant.
+   * */
+  public boolean isCentralTenantContext() {
+    return userTenantsService.getCentralTenant(context.getTenantId())
+      .map(tenantId -> Objects.equals(tenantId, context.getTenantId()))
+      .orElse(false);
   }
 
   private List<String> getTenantsList(List<ConsortiumTenantsClient.ConsortiumTenant> consortiumTenants) {