+++ src/main/java/org/folio/entlinks/utils/ConsortiumUtils.java
@@ -1,0 +1,12 @@
+package org.folio.entlinks.utils;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class ConsortiumUtils {
+  public static final String CONSORTIUM_SOURCE_PREFIX = "CONSORTIUM-";
+
+  public boolean isConsortiumShadowCopy(String source) {
+    return source != null && source.startsWith(CONSORTIUM_SOURCE_PREFIX);
+  }
+}